Add submit error message on Group Curation
If there’s a form error when you click Submit on the /group-curation page, it can be hard to see why the form won’t submit unless you scroll up. This adds an error message next to the Submit button if there are form errors, and the message automatically disappears once all form errors are cleared.