Fix missing caption
Caused by the wrong conditional for displaying the caption in the drop-down section: instead of checking for the existence of the caption, I was testing for the existence of the calculated excerpt.

Also fixed the caption/excerpt not appearing next to the preview icon.

Also fixed the lack of padding in the drop-down area of antibody characterizations.