Make matrix Visualize button dropdown work
The dropdown was getting hidden because it’s an absolutely positioned element, so it doesn’t push the wrapper element down to make room for it. This moves the Visualze button to its own div, and the height of that div is set with an inline style based on the height of the dropdown.