Seem to have fixed the invariant violation error
Caused by multiple renderings of the facets with various this.props changes each time. This caused multiple GET requests for the subfacets, which caused the invariant violation of the results came back for subfacets that have unmounted. Now we check for facet properties to have actually changed before doing GET requests.