Fix collection rendering
Collections are only supposed to display simple values, but files.columns.quality_metrics is an array of objects, and it has to be because we don’t know what kind of quality metric it holds, so we can’t specify every single simple value within every single quality-metric object. This fixes the collections display so that it’ll show a placeholder when it detects an object.

Also getting rid of a controlled/uncontrolled warning in forms, as well as an undefined context function in forms.

Redmine #4755