Fix raw file table bugs
Two separate problems caused this, though at the same point in the code. One was simply that I used the wrong React key for a component rendered within a loop. I was using copy of the parent loop’s key which caused multiple uses of the same key in the inner loop. I changed that to use the file’s @id in the inner loop because that’s unique each time.

The second problem was that rendering an array of <td> components within a conditional expression no longer works in this version of React. I had to replace it with two conditions — one for each former React element.

Redmine #5070