Fix hidden selected term bug
When the user selects a term below he “See More” line and reloads the page, that facet’s “See More” line should be expanded so you can see the selection, but it appears collapsed on reload, hiding the currently selected term. I think this bug has been around a while. This commit fixes that. It was caused by passig the field to `termSelected` instead of the facet it expected. Redmine #5033