Create and destroy charts as needed
I didn’t originally code the awards page to change after display, so I removed the dynamic aspects of the home page when copying the charting code to the award page. Now that we have the genus buttons, the dynamic code — mostly involving creating and destroying charts as the currently selected buttons dictate.