Fix improper use of this.props
We often (and AirBnB forces us to) use stateless components where component properties come from `props` instead of `this.props`. In this place I forgot to convert our old stateful implementaiton with this.props to the new stateless implementation as props. Fixed this by using props instead of this.props in that component. ENC-3426