Got a nested target search running using ES
Whether it’s the _right_ search or not remains to be seen. Next need to modify the front end to interpret the returned JSON, and need to come up with a way for the front end to know whether it’s dealing with an assay matrix or a target matrix. ENCD-3213