ENCD-2704 file filtering (#1943)
* First implementation of tabbed file graph/tables

File graphs and tables now appear tabbed, with either the graph or all the tables showing. The assembly/annotation now exists in its own section controlling both tabs. Next I need to introduce filtering for archived and revoked files. ENCD-2704

* Add additions and subtractions to url function

Added ways to add or subtract statuses to the standard URL for unreleasedFilesUrl. This lets us more easily add a switch to show revoked and archived files in the file table. ENC-2704

* Have basic archive/revoked file filtering

A basic checkbox now filters out/in archived and revoked files. Need to style the checkbox better, and work out what needs to be done for the graph — does it get filtered with this checkbox or not? ENCD-2704

* First demoable commit

Abandoned the idea of using an iOS-style checkbox partly because that’s a very random decision to use, and it takes a *lot* of CSS to do. Ready to start a demo with this. ENCD-2704

* Update globals test for new properties

Changed the property names in globals.unreleasedFilesUrl, so I had to update those property names in this test. ENCD-2704

* Add borders to browser result tabs

The only other place in our code that uses the <TabPanel> component is the browser view. Changed the styling so that it adds a border and looks less weird with the new light gray background. ENCD-2704