Draw without tabs if hideGraph passed
The hideGraph flag determines whether we should draw a graph or not. With the addition of tabs for the file graph and table, I simply drew nothing into the tab contents if hidGraph was used, which just draws something that looks cut off. I now draw no tabs at all if hideGraph is used and draw a standalone file table instead. ENCD-2704