Allow file table info icons to work again
Rearranging the code for file table/graph tabs meant the connection between them that used to exist for bringing up the modal no longer exists, so I had to put that connection back with a fair amount of code rearranging. The ownership of the information now belongs to the parent so that it pass it to both the file table and graph. ENCD-2704