Fix form validation bugs
The defaultValue method of forms.js checked schema.default for truthyness rather than against undefined, so it was returning undefined even if a default value was specified in the schema. I changed it to compare against undefined instead of truthyness. The genetic_modification.reagent_availability had the required fields specification in the wrong place which caused it to always fail form validation, so that’s fixed too. ENCD-3518