Restore two dbxrefPrefixMap entries
Turns out target.js still uses the dbxrefPrefixMap, and I’d forgotten. I think it can probably get rid of all that and just use the default dbxref code, but it was put there fore some reason to begin with, and it’s surprisingly complicated and has no comments. So I think leave it for now, and if we really feel like it, we can tackle deciphering that later. ENCD-3606