ENCD-3606 Cellosaurus (#2013)
* Change Cellosaurus URL pattern to https

Also removed a completely unrelated now-unused reference to a GeneticModification import. VSCode gave a warning on that. ENCD-3606

* Fix dbxref Jest test

For the Cellosaurus protocol change. ENCD-3606

* Remove dbxrefPrefixMap mapping object

Using a completely different mechanism for mapping dbxref prefixes to URLs with ENCD-3489. That left this object unused, but I forgot to remove it. ENCD-3606

* Restore two dbxrefPrefixMap entries

Turns out target.js still uses the dbxrefPrefixMap, and I’d forgotten. I think it can probably get rid of all that and just use the default dbxref code, but it was put there fore some reason to begin with, and it’s surprisingly complicated and has no comments. So I think leave it for now, and if we really feel like it, we can tackle deciphering that later. ENCD-3606