Handle logging-in problem
Logging in while viewing a GM object that has characterizations with documents caused a partial update of the page, and bad search queries going to the server. Solved this by not mutating the original genetic_modification.characterization arrays and their documents arrays, so that the correct search queries go out when you change you logged-in state. ENCD-3766